--- pkg/auth/authorization.go
@@ -1,11 +1,0 @@
-package auth
-
-import "errors"
-
-func AuthorizeAdminUser(session *Session) error {
-	if !session.IsSuperuser {
-		return errors.New("USER_UNAUTHORIZED")
-	}
-
-	return nil
-}
