+++ pkg/handlers/adminapi/api.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/go-openapi/loads"
 
-	"github.com/transcom/mymove/pkg/auth"
+	"github.com/transcom/mymove/pkg/accesscontrols"
 	"github.com/transcom/mymove/pkg/gen/adminapi"
 	adminops "github.com/transcom/mymove/pkg/gen/adminapi/adminoperations"
 	"github.com/transcom/mymove/pkg/handlers"
@@ -29,7 +29,7 @@ func NewAdminAPIHandler(context handlers.HandlerContext) http.Handler {
 	adminAPI.OfficeIndexOfficeUsersHandler = IndexOfficeUsersHandler{
 		HandlerContext:        context,
 		NewQueryFilter:        query.NewQueryFilter,
-		OfficeUserListFetcher: user.NewOfficeUserListFetcher(queryBuilder, auth.AuthorizeAdminUser),
+		OfficeUserListFetcher: user.NewOfficeUserListFetcher(queryBuilder, accesscontrols.AuthorizeAdminUser),
 	}
 
 	return adminAPI.Serve(nil)
